Método para asinc seguindo a plataforma em que está, exclusão de comentários, alteração do nome da variável queda para aq

+++ JavaApp/src/br/edu/ifba/embarcados/javaapp/conector/IComunicacaoSensores.java
@@ -7,7 +7,7 @@ public interface IComunicacaoSensores extends Library{
 	public int iniciar (String porta);
 	public int ler();
 	public char getQueda();
-	public int getAltitude();
+	public float getAltitude();
 	public int finalizar();
